feat: Implementando estrutura controller para rota de turma
